test_system.py
#!/usr/bin/env python3
"""
Quick test script to verify all components work correctly
"""

import sys
import os

# Add current directory to path
sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))

def test_expert_system():
    """Test the expert system without API"""
    print("=" * 60)
    print("TEST 1: Expert System (No API Required)")
    print("=" * 60)
    try:
        from ml_debugging_expert import run_diagnosis
        
        metrics = {
            "train_accuracy": 95,
            "test_accuracy": 68
        }
        
        result = run_diagnosis(metrics)
        
        if "diagnoses" in result and len(result["diagnoses"]) > 0:
            print("✅ Expert system works!")
            print(f"   Diagnosis: {result['diagnoses'][0][:60]}...")
            return True
        else:
            print("❌ Expert system failed - no diagnoses returned")
            return False
            
    except Exception as e:
        print(f"❌ Expert system failed: {e}")
        return False

def test_gemini_integration():
    """Test Gemini API integration"""
    print("\n" + "=" * 60)
    print("TEST 2: Gemini API Integration")
    print("=" * 60)
    try:
        from gemini_integration import GeminiIntegration
        import os
        from dotenv import load_dotenv
        
        load_dotenv()
        api_key = os.getenv("GOOGLE_API_KEY") or os.getenv("GEMINI_API_KEY")
        
        if not api_key:
            print("⚠️  No API key found in .env - skipping API test")
            return True
        
        gemini = GeminiIntegration(api_key)
        
        # Test metric extraction
        query = "My model gets 92% train accuracy but 68% test accuracy"
        metrics = gemini.extract_metrics(query)
        
        if "train_accuracy" in metrics and "test_accuracy" in metrics:
            print("✅ Gemini API works!")
            print(f"   Extracted: train={metrics['train_accuracy']}%, test={metrics['test_accuracy']}%")
            return True
        else:
            print("⚠️  Gemini API returned empty metrics (might be quota/key issue)")
            return True  # Don't fail on API issues
            
    except Exception as e:
        print(f"⚠️  Gemini API test failed: {e}")
        return True  # Don't fail on API issues

def test_imports():
    """Test that all modules can be imported"""
    print("\n" + "=" * 60)
    print("TEST 3: Module Imports")
    print("=" * 60)
    
    modules = [
        "ml_debugging_expert",
        "gemini_integration",
        "main_cli"
    ]
    
    all_ok = True
    for module in modules:
        try:
            __import__(module)
            print(f"✅ {module} imports successfully")
        except Exception as e:
            print(f"❌ {module} failed: {e}")
            all_ok = False
    
    return all_ok

def main():
    """Run all tests"""
    print("\n" + "🧪 ML Debugging Expert System - Test Suite")
    print("=" * 60)
    
    results = []
    
    # Run tests
    results.append(("Imports", test_imports()))
    results.append(("Expert System", test_expert_system()))
    results.append(("Gemini API", test_gemini_integration()))
    
    # Summary
    print("\n" + "=" * 60)
    print("TEST SUMMARY")
    print("=" * 60)
    
    passed = sum(1 for _, result in results if result)
    total = len(results)
    
    for name, result in results:
        status = "✅ PASS" if result else "❌ FAIL"
        print(f"{status} - {name}")
    
    print("\n" + f"Total: {passed}/{total} tests passed")
    
    if passed == total:
        print("\n🎉 All tests passed! Your system is ready to use.")
        print("\nQuick start:")
        print("  python main_cli.py    # Run CLI interface")
        print("  python gui.py         # Run GUI interface")
        return 0
    else:
        print("\n⚠️  Some tests failed. Check the errors above.")
        return 1

if __name__ == "__main__":
    sys.exit(main())
