requirements.txt
experta==1.9.4
google-generativeai>=0.8.0
python-dotenv==1.0.0
tk